LayoutTests/fast/css/vertical-align-length-copy-bug.html

<!DOCTYPE HTML>
<script src="../js/resources/js-test-pre.js"></script>
<style>
#testDiv a, #testDiv span{display:inline-block;border:1px solid black;}
#testDiv a{width:30px;height:30px;text-align:center;line-height:30px;}
#testDiv a span{width:10px;height:10px;opacity:0.9;vertical-align:-1px;}
</style>
<div id=testDiv style="position:absolute">
<a><span id=test></span></a>
</div>
<script>
description("Test for https://bugs.webkit.org/show_bug.cgi?id=71996");
var test = document.getElementById("test");
var style = window.getComputedStyle(test);
shouldBeEqualToString("style.verticalAlign", "-1px");
shouldBe("test.offsetTop", "10");
isSuccessfullyParsed();
</script>
